Addressing reviewer's comment

+++ src/main/scala/edu/ie3/mobsim/model/EvType.scala
@@ -125,7 +125,7 @@ object EvType {
           .getValue
           .doubleValue()
       ),
-      Each(evTypeInput.getCosPhiRated.doubleValue()),
+      Each(evTypeInput.getCosPhiRated),
     )
   }
